+++ src/set4.c
@@ -507,7 +507,7 @@ Set* set_proj(const Set* set, const Tuple* pattern)
    assert(tuple_is_valid(pattern));
 
    dim = tuple_get_dim(pattern);
-   idx = malloc(sizeof(*idx) * dim);
+   idx = malloc(sizeof(*idx) * (size_t)dim);
 
    assert(idx != NULL);
    
@@ -682,7 +682,7 @@ List* set_subsets_list(
       return list;
    }
    
-   counter = malloc(sizeof(*counter) * subset_size);
+   counter = malloc(sizeof(*counter) * (size_t)subset_size);
 
    assert(counter != NULL);
